@@ -18,7 +18,7 @@
     "debounce": "^1.2.0",
     "electron-devtools-installer": "^3.1.0",
     "electron-window-state": "^5.0.3",
-    "nuxt": "^2.13.2",
+    "nuxt": "^2.13.3",
     "pretty-bytes": "^5.3.0",
     "typeface-roboto": "^0.0.75",
     "vuex-module-decorators": "^0.17.0",
@@ -35,13 +35,13 @@
     "cross-env": "^7.0.2",
     "electron": "^8.3.3",
     "electron-builder": "^22.7.0",
-    "eslint": "^7.3.1",
+    "eslint": "^7.4.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.4",
     "husky": "^4.2.5",
-    "node-sass": "^4.14.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.0.5",
+    "sass": "^1.26.9",
     "sass-loader": "^8.0.2",
     "vuetify-loader": "^1.5.0",
     "worker-loader": "^2.0.0"
